+++ User/user/src/main/java/com/betolara1/user/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.betolara1.user.controller;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -8,11 +9,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.data.domain.Page;
 
 import com.betolara1.user.model.User;
 import com.betolara1.user.service.UserService;
+import com.betolara1.user.dto.request.UpdateUserRequest;
 import com.betolara1.user.dto.response.UserDTO;
 import com.betolara1.user.exception.NotFoundException;
 
@@ -26,48 +29,51 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/listAll")
-    public ResponseEntity<Page<UserDTO>> listAll(
+    @GetMapping("/user/getAll")
+    public ResponseEntity<Page<UserDTO>> getAll(
             @RequestParam(value = "page", defaultValue = "0") int page,
-            @RequestParam(value = "size", defaultValue = "10") int size) {
+            @RequestParam(value = "size", defaultValue = "10") int size,
+            @RequestHeader("X-User-Role") String role,
+            @RequestHeader("X-User-Id") Long userId){
 
-        Page<UserDTO> list = userService.findAllUsers(page, size).map(UserDTO::new);
+        if(role.equals("ADMIN")){
+            Page<UserDTO> list = userService.findAllUsers(page, size).map(UserDTO::new);
 
-        if (list.isEmpty()) {
-            throw new NotFoundException("Nenhum usuário cadastrado.");
+            if (list.isEmpty()) {
+                throw new NotFoundException("Nenhum usuário cadastrado.");
+            }
+
+            return ResponseEntity.ok(list);
         }
+        else if(role.equals("USER")){
+            Page<UserDTO> list = userService.findById(userId, page, size).map(UserDTO::new);
 
-        return ResponseEntity.ok(list);
-    }
+            if (list.isEmpty()) {
+                throw new NotFoundException("Nenhum usuário cadastrado.");
+            }
 
-    @GetMapping("/{identifier}")
-    public ResponseEntity<UserDTO> getUser(@PathVariable String identifier) {
-        User user;
-
-        // Verifica se a String contém apenas dígitos (0-9)
-        // Se for o caso, trata como ID do usuário
-        if (identifier.matches("\\d+")) { // Verifica se a string contém apenas dígitos (0-9)
-            Long id = Long.parseLong(identifier);
-            user = userService.findById(id);
-        } else {
-            // Caso contrário, trata como username
-            user = userService.findByUsername(identifier);
+            return ResponseEntity.ok(list);
         }
-
-        return ResponseEntity.ok(new UserDTO(user));
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
-        return ResponseEntity.ok("Usuário deletado com sucesso");
+    @DeleteMapping("/user/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable Long id, @RequestHeader("X-User-Role") String role) {
+        if(role.equals("ADMIN")){
+            userService.deleteUser(id);
+            return ResponseEntity.ok("Usuário deletado com sucesso");
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
-    @PutMapping("/{id}")
-    public ResponseEntity<UserDTO> updateUser(@PathVariable Long id, @RequestBody User user) {
-        User userDB = userService.updateUser(id, user);
-        UserDTO userDTOUpdated = new UserDTO(userDB);
+    @PutMapping("/user/{id}")
+    public ResponseEntity<UserDTO> updateUser(@PathVariable Long id, @RequestBody UpdateUserRequest updatedUser, @RequestHeader("X-User-Id") Long userId) {
+        if(userId.equals(id)){
+            User userDB = userService.updateUser(id, updatedUser);
+            UserDTO userDTOUpdated = new UserDTO(userDB);
 
-        return ResponseEntity.ok(userDTOUpdated);
+            return ResponseEntity.ok(userDTOUpdated);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 }
